+++ app/src/main/java/com/fibelatti/photowidget/platform/ContextKtx.kt
@@ -1,11 +1,15 @@
 package com.fibelatti.photowidget.platform
 
+import android.app.ActivityManager
 import android.content.Context
 import android.content.ContextWrapper
+import android.os.Build
+import android.os.PowerManager
 import androidx.annotation.AttrRes
 import androidx.annotation.ColorInt
 import androidx.annotation.StyleRes
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.getSystemService
 import com.google.android.material.color.DynamicColors
 
 @ColorInt
@@ -36,4 +40,22 @@ fun Context?.findActivity(): AppCompatActivity? {
         currentContext = currentContext.baseContext
     }
     return null
+}
+
+fun Context.isBackgroundRestricted(checkUnrestrictedBattery: Boolean = false): Boolean {
+    val restrictiveManufacturers = listOf("huawei", "xiaomi", "redmi", "oppo", "vivo", "realme", "blackview", "tecno")
+    val isRestrictive = Build.MANUFACTURER.lowercase() in restrictiveManufacturers
+
+    val activityManager: ActivityManager? = getSystemService()
+    val isBackgroundRestricted = Build.VERSION.SDK_INT >= Build.VERSION_CODES.P &&
+        activityManager?.isBackgroundRestricted == true
+
+    val isBatteryUsageRestricted = if (checkUnrestrictedBattery) {
+        val powerManager: PowerManager? = getSystemService()
+        powerManager?.isIgnoringBatteryOptimizations(packageName) != true
+    } else {
+        false
+    }
+
+    return isRestrictive || isBackgroundRestricted || isBatteryUsageRestricted
 }